refactor: ローテータ設定／動作設定／Formのファイルを分割
