+++ src/renderer/components/organisms/setting/RotatorSetting/RotatorBehavior/RotatorBehaviorForm.ts
@@ -1,0 +1,12 @@
+/**
+ * ローテーター設定／動作設定フォーム
+ */
+export default class RotatorBehaviorForm {
+  public basePositionDegree = "";
+  public rangeAzMin = "";
+  public rangeAzMax = "";
+  public moveMode = "";
+  public startAgoMinute = "";
+  public parkPosAz = "";
+  public parkPosEl = "";
+}
