+++ src/renderer/components/organisms/setting/RotatorSetting/RotatorConn/RotatorConnForm.ts
@@ -1,0 +1,13 @@
+/**
+ * ローテーター設定／機器設定フォーム
+ */
+export default class RotatorConnForm {
+  public port = "";
+  public makerId = "";
+  public rotatorId = "";
+  public borate = "";
+  public testAz = "";
+  public testEl = "";
+  public ipAddress = "";
+  public ipPort = "";
+}
